fix: compatibility with discourse version 2.8.7
