+++ assets/javascripts/lib/shorten-number.js
