+++ assets/javascripts/lib/stats/index.js
