+++ assets/javascripts/lib/voting-history/gql/fetcher.js
