+++ assets/javascripts/lib/voting-history/gql/off-chain-fetcher.js
