+++ assets/javascripts/lib/voting-history/gql/on-chain-fetcher.js
