+++ assets/javascripts/lib/voting-history/gql/queries.js
