+++ assets/javascripts/lib/voting-history/index.js
